+++ src/index.ts
@@ -71,3 +71,9 @@ export * from './core/contains';
 export * from './core/containsErr';
 export * from './core/isResult';
 export { matchTag } from './core/matcher';
+export type {
+    AsyncErrMatchBuilder,
+    AsyncErrorMatchBuilder,
+    ErrMatchBuilder,
+    ErrorMatchBuilder,
+} from './core/matcher';